requirements.txt
# 1. django
django>=5.0.1,<6.0.0
# 2. django rest framework
djangorestframework>=3.14.0,<4.0.0
# 3. psycopg2 - postgresql adapter
psycopg2>=2.9.9,<3.0.0
# 4. drf-spectacular : 
drf-spectacular>=0.27.1,<0.28.0
# 5. channels
channels>=4.0.0,<5.0.0